#1134 Move OpenCL support functions generation into the transformation
