+++ src/psyclone/tests/domain/gocean/transformations/gocean_opencl_trans_test.py
@@ -165,9 +165,9 @@ def test_invoke_use_stmts_and_decls(kernel_outputdir, monkeypatch, debug_mode):
         expected += "      use ocl_utils_mod, only: check_status\n"
 
     expected += '''\
-      use fortcl, only: get_num_cmd_queues, get_cmd_queues, get_kernel_by_name
       use clfortran
       use iso_c_binding
+      use fortcl, only: get_num_cmd_queues, get_cmd_queues, get_kernel_by_name
       type(r2d_field), intent(inout), target :: cu_fld, p_fld, u_fld
       integer ystop
       integer ystart
@@ -181,13 +181,11 @@ def test_invoke_use_stmts_and_decls(kernel_outputdir, monkeypatch, debug_mode):
       integer(kind=c_intptr_t) p_fld_cl_mem
       integer(kind=c_intptr_t) cu_fld_cl_mem
       integer(kind=c_intptr_t), target, save :: kernel_compute_cu_code
-      logical, save :: first_time=.true.
+      logical, save :: first_time = .true.
       integer ierr
       integer(kind=c_intptr_t), pointer, save :: cmd_queues(:)
       integer, save :: num_cmd_queues
       '''
-    print(generated_code)
-    assert False
     assert expected in generated_code
     assert GOcean1p0OpenCLBuild(kernel_outputdir).code_compiles(psy)
 
@@ -216,7 +214,7 @@ def test_invoke_opencl_initialisation(kernel_outputdir):
       integer(kind=c_intptr_t) p_fld_cl_mem
       integer(kind=c_intptr_t) cu_fld_cl_mem
       integer(kind=c_intptr_t), target, save :: kernel_compute_cu_code
-      logical, save :: first_time=.true.
+      logical, save :: first_time = .true.
       integer ierr
       integer(kind=c_intptr_t), pointer, save :: cmd_queues(:)
       integer, save :: num_cmd_queues'''
@@ -234,7 +232,6 @@ def test_invoke_opencl_initialisation(kernel_outputdir):
     expected = '''\
       if (first_time) then
         first_time = .false.
-        ! ensure opencl run-time is initialised for this psy-layer module
         call psy_init
         num_cmd_queues = get_num_cmd_queues()
         cmd_queues => get_cmd_queues()
@@ -242,20 +239,21 @@ def test_invoke_opencl_initialisation(kernel_outputdir):
         call initialise_device_buffer(cu_fld)
         call initialise_device_buffer(p_fld)
         call initialise_device_buffer(u_fld)
-        xstart = cu_fld%internal%xstart
-        xstop = cu_fld%internal%xstop
-        ystart = cu_fld%internal%ystart
-        ystop = cu_fld%internal%ystop
-      cu_fld_cl_mem = transfer(cu_fld%device_ptr, cu_fld_cl_mem)
-      p_fld_cl_mem = transfer(p_fld%device_ptr, p_fld_cl_mem)
-      u_fld_cl_mem = transfer(u_fld%device_ptr, u_fld_cl_mem)
+        xstart = cu_fld % internal % xstart
+        xstop = cu_fld % internal % xstop
+        ystart = cu_fld % internal % ystart
+        ystop = cu_fld % internal % ystop
+        cu_fld_cl_mem = transfer(cu_fld % device_ptr, cu_fld_cl_mem)
+        p_fld_cl_mem = transfer(p_fld % device_ptr, p_fld_cl_mem)
+        u_fld_cl_mem = transfer(u_fld % device_ptr, u_fld_cl_mem)
         call compute_cu_code_set_args(kernel_compute_cu_code, cu_fld_cl_mem, \
 p_fld_cl_mem, u_fld_cl_mem, xstart - 1, xstop - 1, ystart - 1, ystop - 1)
-        call cu_fld%write_to_device()
-        call p_fld%write_to_device()
-        call u_fld%write_to_device()
+        call cu_fld % write_to_device
+        call p_fld % write_to_device
+        call u_fld % write_to_device
       end if'''
 
+    print(generated_code)
     assert expected in generated_code
     assert GOcean1p0OpenCLBuild(kernel_outputdir).code_compiles(psy)
 
@@ -340,7 +338,6 @@ c_sizeof(field%grid%area_t(1,1))'''
     expected = '''
       if (first_time) then
         first_time = .false.
-        ! ensure opencl run-time is initialised for this psy-layer module
         call psy_init
         num_cmd_queues = get_num_cmd_queues()
         cmd_queues => get_cmd_queues()
@@ -350,22 +347,22 @@ c_sizeof(field%grid%area_t(1,1))'''
         call initialise_device_buffer(in_fld)
         call initialise_device_buffer(dx)
         call initialise_grid_device_buffers(in_fld)
-        xstart = out_fld%internal%xstart
-        xstop = out_fld%internal%xstop
-        ystart = out_fld%internal%ystart
-        ystop = out_fld%internal%ystop
-      out_fld_cl_mem = transfer(out_fld%device_ptr, out_fld_cl_mem)
-      in_out_fld_cl_mem = transfer(in_out_fld%device_ptr, in_out_fld_cl_mem)
-      in_fld_cl_mem = transfer(in_fld%device_ptr, in_fld_cl_mem)
-      dx_cl_mem = transfer(dx%device_ptr, dx_cl_mem)
-      gphiu_cl_mem = transfer(in_fld%grid%gphiu_device, gphiu_cl_mem)
+        xstart = out_fld % internal % xstart
+        xstop = out_fld % internal % xstop
+        ystart = out_fld % internal % ystart
+        ystop = out_fld % internal % ystop
+        out_fld_cl_mem = transfer(out_fld % device_ptr, out_fld_cl_mem)
+        in_out_fld_cl_mem = transfer(in_out_fld % device_ptr, in_out_fld_cl_mem)
+        in_fld_cl_mem = transfer(in_fld % device_ptr, in_fld_cl_mem)
+        dx_cl_mem = transfer(dx % device_ptr, dx_cl_mem)
+        gphiu_cl_mem = transfer(in_fld % grid % gphiu_device, gphiu_cl_mem)
         call compute_kernel_code_set_args(kernel_compute_kernel_code, \
-out_fld_cl_mem, in_out_fld_cl_mem, in_fld_cl_mem, dx_cl_mem, in_fld%grid%dx, \
+out_fld_cl_mem, in_out_fld_cl_mem, in_fld_cl_mem, dx_cl_mem, in_fld % grid % dx, \
 gphiu_cl_mem, xstart - 1, xstop - 1, ystart - 1, ystop - 1)
-        call out_fld%write_to_device()
-        call in_out_fld%write_to_device()
-        call in_fld%write_to_device()
-        call dx%write_to_device()
+        call out_fld % write_to_device
+        call in_out_fld % write_to_device
+        call in_fld % write_to_device
+        call dx % write_to_device
         call write_grid_buffers(in_fld)
       end if'''
     assert expected in generated_code
@@ -521,50 +518,43 @@ def test_psy_init(kernel_outputdir, monkeypatch):
     otrans = GOOpenCLTrans()
     otrans.apply(sched)
     generated_code = str(psy.gen)
-    expected = (
-        "    SUBROUTINE psy_init()\n"
-        "      USE fortcl, ONLY: ocl_env_init, add_kernels\n"
-        "      CHARACTER(LEN=30) kernel_names(1)\n"
-        "      INTEGER :: ocl_device_num=1\n"
-        "      LOGICAL, save :: initialised=.False.\n"
-        "      ! Check to make sure we only execute this routine once\n"
-        "      IF (.not. initialised) THEN\n"
-        "        initialised = .True.\n"
-        "        ! Initialise the OpenCL environment/device\n"
-        "        CALL ocl_env_init(1, ocl_device_num, .False., .False.)\n"
-        "        ! The kernels this PSy layer module requires\n"
-        "        kernel_names(1) = \"compute_cu_code\"\n"
-        "        ! Create the OpenCL kernel objects. Expects to find all of "
-        "the compiled\n"
-        "        ! kernels in FORTCL_KERNELS_FILE.\n"
-        "        CALL add_kernels(1, kernel_names)\n"
-        "      END IF\n"
-        "    END SUBROUTINE psy_init\n")
+    expected = '''
+    SUBROUTINE psy_init()
+      USE fortcl, ONLY: add_kernels, ocl_env_init
+      CHARACTER(LEN=30) kernel_names(1)
+      INTEGER :: ocl_device_num = 1
+      LOGICAL, SAVE :: initialised = .FALSE.
+
+      IF (.NOT.initialised) THEN
+        initialised = .true.
+        CALL ocl_env_init(1, ocl_device_num, .false., .false.)
+        kernel_names(1) = 'compute_cu_code'
+        CALL add_kernels(1, kernel_names)
+      END IF
+
+    END SUBROUTINE psy_init'''
     assert expected in generated_code
     assert GOcean1p0OpenCLBuild(kernel_outputdir).code_compiles(psy)
 
     # Test with a non-default number of OpenCL queues
     sched.coded_kernels()[0].set_opencl_options({'queue_number': 5})
     generated_code = str(psy.gen)
-    expected = (
-        "    SUBROUTINE psy_init()\n"
-        "      USE fortcl, ONLY: ocl_env_init, add_kernels\n"
-        "      CHARACTER(LEN=30) kernel_names(1)\n"
-        "      INTEGER :: ocl_device_num=1\n"
-        "      LOGICAL, save :: initialised=.False.\n"
-        "      ! Check to make sure we only execute this routine once\n"
-        "      IF (.not. initialised) THEN\n"
-        "        initialised = .True.\n"
-        "        ! Initialise the OpenCL environment/device\n"
-        "        CALL ocl_env_init(5, ocl_device_num, .False., .False.)\n"
-        "        ! The kernels this PSy layer module requires\n"
-        "        kernel_names(1) = \"compute_cu_code\"\n"
-        "        ! Create the OpenCL kernel objects. Expects to find all of "
-        "the compiled\n"
-        "        ! kernels in FORTCL_KERNELS_FILE.\n"
-        "        CALL add_kernels(1, kernel_names)\n"
-        "      END IF\n"
-        "    END SUBROUTINE psy_init\n")
+    expected = '''
+    SUBROUTINE psy_init()
+      USE fortcl, ONLY: add_kernels, ocl_env_init
+      CHARACTER(LEN=30) kernel_names(1)
+      INTEGER :: ocl_device_num = 1
+      LOGICAL, SAVE :: initialised = .FALSE.
+
+      IF (.NOT.initialised) THEN
+        initialised = .true.
+        CALL ocl_env_init(1, ocl_device_num, .false., .false.)
+        kernel_names(1) = 'compute_cu_code'
+        CALL add_kernels(1, kernel_names)
+      END IF
+
+    END SUBROUTINE psy_init'''
+    print(generated_code)
     assert expected in generated_code
     assert GOcean1p0OpenCLBuild(kernel_outputdir).code_compiles(psy)
